Reformat RefundModelTest
